[FP-215] fix : 경기 - 경기장 캐싱 오류 수정
